==> 03-household_cond_data.Rmd <==
---
title: "ENIGH Data Analysis"
author: "Mauricio Hernandez"
date: "2/22/2021"
subtitle: Organizing Household Condensed Data
output:
  html_document:
    df_print: paged
---

```{r, include=FALSE}
#Remove all objects from current workspace and call garbage collector
rm(list=ls())
gc()

#Set WD
current_path = rstudioapi::getActiveDocumentContext()$path 
setwd(dirname(current_path ))
```

```{r sh, include=FALSE}
source("./script/setup_functions.R")
```


```{r knitr_init, include=FALSE}
#options(max.print = "75")
opts_chunk$set(echo = TRUE,
	             cache = FALSE,
               prompt = FALSE,
               tidy = FALSE,
               comment = NA,
               message = FALSE,
               warning = FALSE,
               fig.path = 'figs/')
```

```{r, include=FALSE}
#Generates the css needed by summarytools
st_css()
```

```{r setup, include=FALSE}
knitr::opts_chunk$set(fig.width = 8, collapse = TRUE)

knitr::knit_hooks$set(output = function(x, options){
  if(!is.null(options$max_height)){
    paste('<pre style = "max-height:',
          options$max_height, 
          ';float: left; width: 910px; overflow-y: auto;">',
          x,
          "</pre>",
          sep = "")
  } else{
    x
  }
})


#Set WD
current_path = rstudioapi::getActiveDocumentContext()$path 
setwd(dirname(current_path ))
print( getwd() )

#Load Packages
if (!require("pacman")) install.packages("pacman")
  pacman::p_load(pacman, party, psych, rio, tidyverse, knitr, bit64, plotly, 
                 forcats, stargazer, summarytools)

opts_chunk$set(tidy.opts=list(width.cutoff=60),tidy=TRUE)
```

## Reading ENIGH microdata (CSV files)
```{r include = TRUE, echo = FALSE}
path.data <- "https://mauricioh2.com/disreu/inegi/enigh/"


#df_hsd_condens <-read_csv(paste0(path.data, "2018_enigh_viviendas.csv"), col_names = TRUE,
#                    col_types=c("folioviv"="c", 'estufa_chi'='i'), na ='&')

df_hsd_condens <-read_csv(paste0(path.data, "2018_enigh_concentradohogar.csv"), col_names = TRUE,
                   col_types=c("folioviv"="c"), na ='&')

```

```{r include = TRUE, echo = FALSE,  results='hide'}
df_hsd_exp <-read_csv("./inputs/2018_enigh_gastoshogar.csv", 
                           col_names = TRUE, progress = TRUE, na ='-1',
                   col_types=cols_only("folioviv"='c', "foliohog"='c', "clave"='c',
                                       "tipo_gasto"='c', "mes_dia"='c', "forma_pag1"='i', 
                                       "forma_pag2"='i', "forma_pag3"='i', "lugar_comp"='i',
                                       "orga_inst"='i', "frecuencia"='i', "fecha_adqu"='c', 
                                       "fecha_pago"='c', "cantidad"='d', "gasto"='d',
                                       "pago_mp"='d', "costo"='d', "inmujer"='d',
                                       "inst_1"='i', "inst_2"='i', "num_meses"='i',
                                       "num_pagos"='i', "ultim_pago"='c', "gasto_tri"='d',
                                       "gasto_nm"='d', "gas_nm_tri"='d', "imujer_tri"='d'),)
```


```{r}
df_person_exp <-read_csv(paste0(path.data, "2018_enigh_gastospersona.csv"),
                              col_names = TRUE,
                   col_types=c("folioviv"="c", 'inst' ='i', 'colegia' = 'd', 
                               'material' = 'd', 'inscrip' = 'd'), na ='-1')
```

## Descriptive Statistics {.tabset}

### Households Condensed

```{r}
descr(df_hsd_condens, stats = c("mean", "sd", "min", "med", "max", "n.valid", "pct.valid"), 
      transpose = TRUE)
```

### Households Expenses

```{r}
descr(df_hsd_exp, stats = c("mean", "sd", "min", "med", "max", "n.valid", "pct.valid"), 
      transpose = TRUE)
```

### People Expenses
```{r}
descr(df_person_exp, stats = c("mean", "sd", "min", "med", "max", "n.valid", "pct.valid"), 
      transpose = TRUE)
```
## Check Energy Expenses

Check all the energy expenses in the households expenses data set and people expenses data sets are the same than the ones reported in the households condensed data set.

df_hsd_condens$energia is equal to the sum of gastoshogar.gasto_tri if clave is equal to G009-G016, R001, R003 plus 
gastospersona.gasto_tri if clave is equal to G009-G016, R001, R003.


|Code (clave) |    Description         |
| ----------- |:----------------------:|
|G009    	    |Liquefied petroleum gas |
|G010    	    |Petroleum               |
|G011    	    |Diesel                  |
|G012    	    |Carbon                  |
|G013    	    |Firewood                |
|G014    	    |Fuel to heat            |
|G015    	    |Candles                 |
|G016    	    |Other fuels             |
|R001    	    |Electricity             |
|R003    	    |Natural Gas			       |


```{r}
df_hsd_exp_energy <- subset(df_hsd_exp, clave=="G009" | clave=="G010" | 
                                   clave=="G011" | clave=="G012" | clave=="G013" | 
                                   clave=="G014" | clave=="G015" | clave=="G016" | 
                                   clave =='R001' | clave =='R003')

df_person_exp_energy <- subset(df_person_exp, clave=="G009" | clave=="G010" | 
                                   clave=="G011" | clave=="G012" | clave=="G013" | 
                                   clave=="G014" | clave=="G015" | clave=="G016" | 
                                   clave =='R001' | clave =='R003')
```


```{r}

write.csv(df_hsd_exp_energy, file = 'outputs/2018_enigh_hsd_energy_exp.csv',
          na = "NULL", fileEncoding="UTF-8", row.names=FALSE)

write.csv(df_person_exp_energy, file = 'outputs/2018_enigh_psn_energy_exp.csv',
          na = "NULL", fileEncoding="UTF-8", row.names=FALSE)
```


Merging data sets by folioviv using energy expenses related data
```{r}
library(dplyr)

hsd_expenses_by_folio <- df_hsd_exp_energy %>%
  group_by(folioviv) %>% replace(is.na(.), 0) %>% 
  summarise(gasto_tri = sum(gasto_tri)
)

person_expenses_by_folio <- df_person_exp_energy %>%
  group_by(folioviv) %>% replace(is.na(.), 0) %>% 
  summarise(gasto_tri = sum(gasto_tri)
)


df_hsd_condens_energy <- subset(df_hsd_condens, select=c('folioviv','energia')) %>%
  group_by(folioviv) %>% replace(is.na(.), 0) %>% 
  summarise(energia = sum(energia))
       
df_merge <- merge(df_hsd_condens_energy, person_expenses_by_folio,by="folioviv", 
                  all.x = TRUE) %>% 
  replace(is.na(.), 0)

df_merge <- merge(df_merge, hsd_expenses_by_folio, by="folioviv", all.x = TRUE) %>% 
  replace(is.na(.), 0)

df_merge$sum_gastotri <- df_merge$gasto_tri.x + df_merge$gasto_tri.y
df_merge$diff <- df_merge$sum_gastotri - df_merge$energia
df_merge$diff[abs(df_merge$diff) <= 0.0001] <- 0 

df_merge
```

```{r}
descr(df_merge, stats = c("mean", "sd", "min", "med", "max", "n.valid", "pct.valid"), 
      transpose = TRUE)
```


Checking electricity Expenses
```{r}
df_hsd_exp$id_household <- paste0(df_hsd_exp$folioviv, df_hsd_exp$foliohog)

df_person_exp$id_household <- paste0(df_person_exp$folioviv, df_person_exp$foliohog)

df_hsd_exp_electricity<- subset(df_hsd_exp, clave =='R001')
df_person_exp_electricity <- subset(df_person_exp,clave =='R001')
```


```{r plt-bill1-days, fig.cap="Energy Expenses", warning=FALSE, message=FALSE, results="hold"}

df_dummy <- subset(df_hsd_exp_energy, select=c('clave','gasto_tri')) %>%
  group_by(clave) %>% replace(is.na(.), 0) %>% 
  summarise(gasto_tri = mean(gasto_tri))

df_dummy$clave[df_dummy$clave=='G009'] <- 'LPG'
df_dummy$clave[df_dummy$clave=='G010'] <- 'Petroleum'
df_dummy$clave[df_dummy$clave=='G011'] <- 'Diesel'
df_dummy$clave[df_dummy$clave=='G012'] <- 'Carbon'
df_dummy$clave[df_dummy$clave=='G013'] <- 'Firewood'
df_dummy$clave[df_dummy$clave=='G014'] <- 'Fuel to heat'
df_dummy$clave[df_dummy$clave=='G015'] <- 'Candles'
df_dummy$clave[df_dummy$clave=='G016'] <- 'Other fuels'
df_dummy$clave[df_dummy$clave=='R001'] <- 'Electricity'
df_dummy$clave[df_dummy$clave=='R003'] <- 'Natural Gas'

pl_enegy_exp <- plot_ly(x = df_dummy$clave, 
                        y = df_dummy$gasto_tri, type = "bar") %>%
  layout(title = 'Average Quarterly Expenses in Energy per Household ($MXN)',
         xaxis = list(title = "Type of Energy"), 
         yaxis = list(title = "$MXN"))

pl_enegy_exp

plotly_IMAGE(pl_enegy_exp, format = "png", out_file = "./figs/energy_expenses.png" )
```
